+++ CHANGELOG.md
@@ -14,6 +14,10 @@ use patch releases for compatibility fixes instead.
 - Added `Beaker.workspace.iter_(images|experiments|datasets)` methods.
 - Added `older_than` parameter to `Beaker.workspace.clear()` method.
 
+### Fixed
+
+- Fixed `Beaker.experiment.results()` when dataset no longer exists. It will now return `None` instead of failing with `DatasetNotFound`.
+
 ## [v1.11.6](https://github.com/allenai/beaker-py/releases/tag/v1.11.6) - 2022-11-22
 
 ### Added